requirements.test.txt
# For testing requirements
pytest
pytest-cov
# target HA 2021.11.5
pytest-homeassistant-custom-component==0.4.12

# Match HA 2021.11.5 pre_commit requirements.
black==21.9b0
codespell==2.0.0
flake8-comprehensions==3.7.0
flake8-docstrings==1.6.0
flake8-noqa==1.2.0
flake8==4.0.1
isort==5.9.3
mccabe==0.6.1
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pyupgrade==2.27.0

# Local packages installed prior to requirements file
mypy==0.910
pylint==2.10.2
